atualizado permissoes e adicionado validacoes ao script do ionic

+++ ubuntu_ionic_installer.sh
@@ -29,14 +29,17 @@ ANDROID_SDK="https://dl.google.com/android/android-sdk_r24.0.2-linux.tgz"
 #Desktop is default in linux English version, but in another language that folder not exist
 cd ~/
 
-wget "$ANDROID_SDK" -O "android-sdk.tgz"
+if [ ! -f android-sdk.tgz ]; then
+    echo "File not found!"
+	wget "$ANDROID_SDK" -O "android-sdk.tgz"
+fi
 
-tar zxf "android-sdk.tgz" -C "$INSTALL_PATH"
-
-cd "$INSTALL_PATH" &&  mv "android-sdk-linux" "android-sdk"
-
+if [ ! -d android-sdk ]; then
+	tar zxf "android-sdk.tgz" -C "$INSTALL_PATH"
+	cd "$INSTALL_PATH" &&  mv "android-sdk-linux" "android-sdk"
+fi
 # Android SDK requires some x86 architecture libraries even on x64 system
-apt-get install -qq -y libc6:i386 libgcc1:i386 libstdc++6:i386 libz1:i386
+apt-get install -y libc6:i386 libgcc1:i386 libstdc++6:i386 libz1:i386
 
 cd "$INSTALL_PATH" &&  chown root:root "android-sdk" -R
 cd "$INSTALL_PATH" &&  chmod 777 "android-sdk" -R
@@ -48,7 +51,7 @@ echo "export PATH=\$PATH:$ANDROID_SDK_PATH/tools" >> ".profile"
 echo "export PATH=\$PATH:$ANDROID_SDK_PATH/platform-tools" >> ".profile"
 
 # Install JDK and Apache Ant
-apt-get -qq -y install default-jdk ant
+apt-get -y install default-jdk ant
 
 # Set JAVA_HOME based on the default OpenJDK installed
 export JAVA_HOME="$(find /usr -type l -name 'default-java')"
@@ -56,7 +59,9 @@ if [ "$JAVA_HOME" != "" ]; then
     echo "export JAVA_HOME=$JAVA_HOME" >> ".profile"
 fi
 
-curl -sL https://deb.nodesource.com/setup | sudo bash -
+echo "Instalando NODEJS 4.x"
+apt-get -y install curl
+curl -sL https://deb.nodesource.com/setup_4.x | sudo bash -
 apt-get update
 apt-get install -y nodejs
